refactored some functions to new utils module. fixed formatting
